More fixes concerning Animation easing.
